+++ coreNetwork/src/main/kotlin/io/linka/app/kotlin/core/network/CoreNetworkModulo.kt
@@ -6,4 +6,8 @@ object CoreNetworkModulo {
     fun criarMonitorRede(context: Context): MonitorRede {
         return MonitorRedeAndroid(context)
     }
+
+    fun criarNetworkCapabilitiesProvider(context: Context): NetworkCapabilitiesProvider {
+        return NetworkCapabilitiesProviderImpl(context)
+    }
 }